slub: add support for verifying slab sanitization
This could support !PAGE_SANITIZE by zeroing in the object setup code in
those cases, but for now it takes the simple approach of depending on
page sanitization.

This is an extension to the sanitization feature in PaX for when
sacricifing more performance for security is acceptable.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>

Cherry-picked from branch '4.17'
of https://github.com/anthraxx/linux-hardened.git
Change-Id: I881fccff1e7a28ce9f5276a8cafe848f22068be9