x86_64: move vdso to mmap region from stack region
This removes the only executable code from the stack region and gives
the vdso the same randomized base as other mmap mappings including the
linker and other shared objects. It results in a sane amount of entropy
being provided and there's little to no advantage in separating this
from the existing executable code there.

It's sensible for userspace to reserve the initial mmap base as a region
for executable code with a random gap for other mmap allocations, along
with providing randomization within that region. However, there isn't
much the kernel can do to help due to how dynamic linkers load the
shared objects.

This was extracted from the PaX RANDMMAP feature.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>

Cherry-picked from branch '4.17'
of https://github.com/anthraxx/linux-hardened.git
Change-Id: I8878cdf994b1bd50413c2cb1eee895e795a84e47