Lock down TIOCSSERIAL
Lock down TIOCSSERIAL as that can be used to change the ioport and irq
settings on a serial port.  This only appears to be an issue for the serial
drivers that use the core serial code.  All other drivers seem to either
ignore attempts to change port/irq or give an error.

Reported-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: David Howells <dhowells@redhat.com>
cc: Jiri Slaby <jslaby@suse.com>

Cherry-picked from branch 'lockdown'
of https://git.kernel.org/pub/scm/linux/kernel/git/jforbes/linux.git
Change-Id: If033edc6ac97c7062579a9fa5be31d7cb4412fe5